third_party/WebKit/LayoutTests/inspector/network/network-toggle-type-filter.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>
function test() {
    var panel = WebInspector.inspectorView._showPanel("network");
    InspectorTest.recordNetwork();
    var target = panel._networkLogView;
    var types = WebInspector.resourceTypes;

    function toggleAndDump(buttonName, toggle)
    {
        InspectorTest.addResult("");
        InspectorTest.addResult((toggle ? "Toggled '" : "Clicked '") + buttonName + "' button.");
        target._resourceTypeFilterUI._toggleTypeFilter(buttonName, toggle);
        var results = [];
        var request = new WebInspector.NetworkRequest(WebInspector.targetManager.mainTarget(), "", "", "", "", "");
        for (var typeId in types) {
            var type = WebInspector.resourceTypes[typeId];
            results.push(type.name() + ": " + target._resourceTypeFilterUI.accept(type.name()));
        }
        InspectorTest.addResult("Filter: " + results.join(", "));
    }

    var allTypes = "all";

    toggleAndDump(allTypes, false);
    toggleAndDump(types.Document.name(), false);
    toggleAndDump(types.Document.name(), false);
    toggleAndDump(types.Script.name(), false);

    toggleAndDump(allTypes, true);
    toggleAndDump(allTypes, true);
    toggleAndDump(types.Stylesheet.name(), true);
    toggleAndDump(types.Image.name(), true);
    toggleAndDump(types.Stylesheet.name(), true);
    toggleAndDump(types.XHR.name(), false);

    toggleAndDump(types.Font.name(), true);
    toggleAndDump(types.WebSocket.name(), true);
    toggleAndDump(types.Media.name(), true);
    toggleAndDump(types.TextTrack.name(), true);
    toggleAndDump(allTypes, false);

    InspectorTest.completeTest();
}

</script>
</head>
<body onload="runTest()">
<p>Tests toggling type filter on network panel.</p>
</body>
</html>
